Fix: Bug with Setting Drink Strength
Bug description: drinkStrength for calculations depends on the last input, not on the mode of setting. Thus for instance, if a user selects a drink when he chooses to set alcohol content manually, it's the selected drink's strength that will be calculated, not the entered manually one.

Solution: use useEffect hook to detect when one of the three options changes (dropdown, input or mode of setting) and update drinkStrength with the handleDrinkStrengthChange function depending on the mode a user chooses. Also, the function fires only in the useEffect to prevent unnecessary changes.